Restore example build scripts.
