+++ examples/testAll.js
