+++ package.json
@@ -23,7 +23,7 @@
     "test:cov": "cross-env BABEL_ENV=commonjs babel-node $(npm bin)/isparta cover $(npm bin)/_mocha -- --recursive",
     "test:examples": "cross-env BABEL_ENV=commonjs babel-node examples/testAll.js",
     "check:src": "npm run lint:src && npm run test",
-    "check:examples": "npm run build:examples && npm run test:examples",
+    "check:examples": "npm run build:examples && npm run lint:examples && npm run test:examples",
     "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir lib",
     "build:es": "cross-env BABEL_ENV=es babel src --out-dir es",
     "build:umd": "cross-env BABEL_ENV=commonjs NODE_ENV=development webpack src/index.js dist/redux.js",
